lxd/firewall: Updates interface and loader for new pkg
